added copyright, commented fields and methods, addressed feedback.
