+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/download/http/HttpConnectionReleaseDownloader.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
 package com.microsoft.appcenter.distribute.download.http;
 
 import android.Manifest;
@@ -5,8 +10,7 @@ import android.content.Context;
 import android.os.Build;
 import android.support.annotation.NonNull;
 
-import com.microsoft.appcenter.distribute.InstallerUtils;
-import com.microsoft.appcenter.distribute.PermissionsUtil;
+import com.microsoft.appcenter.distribute.PermissionUtils;
 import com.microsoft.appcenter.distribute.ReleaseDetails;
 import com.microsoft.appcenter.distribute.download.ReleaseDownloader;
 import com.microsoft.appcenter.utils.AsyncTaskUtils;
@@ -16,13 +20,13 @@ import java.util.ArrayList;
 
 import static com.microsoft.appcenter.distribute.DistributeConstants.LOG_TAG;
 
-// TODO JavaDoc
+/**
+ * Downloads new releases on Android SDK versions prior to Lollipop.
+ * Uses HttpConnection and AsyncTasks.
+ */
 public class HttpConnectionReleaseDownloader implements ReleaseDownloader {
 
     private Context mContext;
-
-    // TODO Move it to const
-    static final String PREFERENCE_KEY_DOWNLOADED_FILE = "PREFERENCE_KEY_DOWNLOADED_FILE";
 
     public HttpConnectionReleaseDownloader(Context context) {
         mContext = context;
@@ -33,7 +37,7 @@ public class HttpConnectionReleaseDownloader implements ReleaseDownloader {
         if (!prepareDownload(listener)) {
             return;
         }
-        AsyncTaskUtils.execute(LOG_TAG, new HttpDownloadFileTask(releaseDetails, listener));
+        AsyncTaskUtils.execute(LOG_TAG, new HttpDownloadFileTask(releaseDetails, listener, mContext));
     }
 
     @Override
@@ -47,15 +51,9 @@ public class HttpConnectionReleaseDownloader implements ReleaseDownloader {
             return false;
         }
         String[] permissions = requiredPermissions();
-        int[] permissionsState = PermissionsUtil.permissionsState(mContext, permissions);
-        if (!PermissionsUtil.permissionsAreGranted(permissionsState)) {
+        int[] permissionsState = PermissionUtils.permissionsState(mContext, permissions);
+        if (!PermissionUtils.permissionsAreGranted(permissionsState)) {
             listener.onError("No external storage permission.");
-            return false;
-        }
-
-        // TODO it should be already enabled
-        if (!InstallerUtils.isUnknownSourcesEnabled(mContext)) {
-            listener.onError("Install from unknown sources disabled.");
             return false;
         }
         return true;
